Rename `helper.rs` to `die.rs`
`helper.rs` is the file where the Die trait was stored,
now it was moved to a new `die.rs`.
This new file is also registered as a module.
