+++ src/modules/git.rs
@@ -2,7 +2,7 @@ use std::path;
 use std::process::Command;
 
 use crate::core::args::Flags;
-use crate::helper::Die;
+use crate::die::Die;
 
 pub fn changed(file_path: &path::Path, flags: &Flags) -> bool {
   let result = Command::new("git").arg("status").arg("--porcelain").arg(file_path).output();
